several fixes to SunCC build on linux

+++ src/engine/bugengine/src/main-p=windows.cc
@@ -21,19 +21,19 @@ int WINAPI WinMain( HINSTANCE /*hInstance*/,
     //AllocConsole();
     if (GetConsoleWindow())
     {
-        long stdHandle = (long)GetStdHandle(STD_OUTPUT_HANDLE);
+        long stdHandle = static_cast<long>((uintptr_t)GetStdHandle(STD_OUTPUT_HANDLE));
         int conHandle = _open_osfhandle(stdHandle, _O_TEXT);
         FILE* fp = _fdopen(conHandle, "w");
         *stdout = *fp;
         setvbuf( stdout, NULL, _IONBF, 0 );
 
-        stdHandle = (long)GetStdHandle(STD_INPUT_HANDLE);
+        stdHandle = static_cast<long>((uintptr_t)GetStdHandle(STD_INPUT_HANDLE));
         conHandle = _open_osfhandle(stdHandle, _O_TEXT);
         fp = _fdopen( conHandle, "r" );
         *stdin = *fp;
         setvbuf( stdin, NULL, _IONBF, 0 );
 
-        stdHandle = (long)GetStdHandle(STD_ERROR_HANDLE);
+        stdHandle = static_cast<long>((uintptr_t)GetStdHandle(STD_ERROR_HANDLE));
         conHandle = _open_osfhandle(stdHandle, _O_TEXT);
         fp = _fdopen( conHandle, "w" );
         *stderr = *fp;
